pyproject.toml
[project]
name = "sou"
version = "0.0.1"
description = ""
authors = [
  { name="Yulin Shi", email="eliane.shiyulin@gmail.com" },
  { name="Baptiste Aubertin", email="bapt120@hotmail.fr" }
]
dependencies = [
  "litellm",
  "tavily-python",
  "nano_graphrag",
  "docker",
  "torch",
  "transformers"

]

readme = "README.md"
requires-python = ">=3.7"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = [
  "/sou",
]



[project.urls]
"Homepage" = "https://github.com/yourusername/mypackage"
